pyproject.toml
[project]
name = "livekit-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Shakir Ali",email = "shakirali9749@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "livekit-server-sdk-python @ git+https://github.com/livekit/livekit-server-sdk-python.git"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
